改进影片关系分析脚本

+++ topN_tv&cast.py
@@ -19,6 +19,7 @@ parser.add_argument("-p", "--password", help="数据库密码", default="neo4j")
 parser.add_argument("-v", "--verbose", help="显示日志信息", action="store_true")
 parser.add_argument("--thread", help="连接的线程数", type=int, default=8)
 parser.add_argument("--retry", help="连接失败重传数", type=int, default=3)
+parser.add_argument("--clearDB", help="是否在初始化时清空数据库原有内容", action="store_true")
 
 
 def insert_if_not_exists(graph: Graph, id, label, **kwargs)->Node:
@@ -54,14 +55,86 @@ def download_credit_if_not_exists(graph: Graph, api: TMDBApi, id, label)->Node:
         return nodes.first()
 
 
+def import_credit_info(graph, api, id, label, movie_node):
+    # 导入演员关系
+    logging.info(f"导入演员关系...")
+    credits = api.get_credits(id, label)  # 获取电影的演职员表
+    for title, credit_list in credits.items():
+        if title == "id":
+            continue
+        for credit in credit_list:
+            character = credit.get("character") or credit.get("job")
+            order = credit.get("order", 0)
+            credit_id = credit["id"]
+            # 插入或搜索到节点
+            credit_node = download_credit_if_not_exists(
+                graph, api, credit_id, title)
+            # 新增一条演员到电影的连线
+            credit_to_movie = Relationship(
+                movie_node, title, credit_node, character=character, order=order)
+            graph.create(credit_to_movie)
+
+
+def import_genre_info(graph, genre_ids, movie_node):
+    logging.info(f"导入类型关系...")
+    for genre_id in genre_ids:
+        genre_nodes = graph.nodes.match("genre", id=genre_id)
+        if not len(genre_nodes):
+            logging.error(f"cannot find genre {genre_id}!")
+        else:
+            genre_node = genre_nodes.first()
+            movie_to_genre = Relationship(movie_node, "is", genre_node)
+            graph.create(movie_to_genre)
+
+
+def import_country_info(graph, country_names, node):
+    logging.info("导入国家关系...")
+    for country_name in country_names:
+        country_node = insert_if_not_exists(graph, country_name, "country")
+        movie_to_country = Relationship(node, "origin_country", country_node)
+        graph.create(movie_to_country)
+
+
+def import_company_info(graph, companies, movie_node):
+    logging.info("导入出品公司信息...")
+    for company in companies:
+        id = company["id"]
+        name = company["name"]
+        country = company.get("origin_country", None)
+        company_node = insert_if_not_exists(graph, id, "company", name=name)
+        import_country_info(graph, [country, ], company_node)
+        movie_to_company = Relationship(
+            movie_node, "production_company", company_node)
+        graph.create(movie_to_company)
+
+
+def import_created_by_info(graph, people, movie_node):
+    logging.info("导入出品人信息...")
+    for person in people:
+        id = person["id"]
+        name = person["name"]
+        gender = person["gender"]
+        person_node = insert_if_not_exists(
+            graph, id, "producer", gender=gender, name=name)
+        movie_to_person = Relationship(movie_node, "created_by", person_node)
+        graph.create(movie_to_person)
+
+
 if __name__ == "__main__":
+
     def process_info(info):
+        id = info.get("id")  # 确定电影id
+        info = api.get_details(id, arg.target)
         name = info.get("name") or info.get("title")  # 确定电影名
         logging.info(f"当前影片: {name}")
-        id = info.get("id")  # 确定电影id
+
         popularity = info.get("popularity", 0)
         vote_average = info.get("vote_average", 0)
-        genre_ids = info.get("genre_ids", [])
+        genre_ids = info.get("genre_ids", []) or (genre["id"]
+                                                  for genre in info.get("genres", []))
+        origin_countries = info.get("origin_country", [])
+        companies = info.get("production_companies", [])
+        created_by = info.get("created_by", [])
 
         # 插入影片节点
         logging.info(f"尝试插入影片节点...")
@@ -74,34 +147,18 @@ if __name__ == "__main__":
         )
 
         # 导入演员关系
-        logging.info(f"导入演员关系...")
-        credits = api.get_credits(id, arg.target)  # 获取电影的演职员表
-        logging.debug(credits)
-        for title, credit_list in credits.items():
-            if title == "id":
-                continue
-            for credit in credit_list:
-                character = credit.get("character") or credit.get("job")
-                order = credit.get("order", 0)
-                credit_id = credit["id"]
-                # 插入或搜索到节点
-                credit_node = download_credit_if_not_exists(
-                    graph, api, credit_id, title)
-                # 新增一条演员到电影的连线
-                credit_to_movie = Relationship(
-                    movie_node, title, credit_node, character=character, order=order)
-                graph.create(credit_to_movie)
+        import_credit_info(graph, api, id, arg.target, movie_node)
 
         # 导入类型关系
-        logging.info(f"导入类型关系...")
-        for genre_id in genre_ids:
-            genre_nodes = graph.nodes.match("genre", id=genre_id)
-            if not len(genre_nodes):
-                logging.error(f"cannot find genre {genre_id}!")
-            else:
-                genre_node = genre_nodes.first()
-                movie_to_genre = Relationship(movie_node, "is", genre_node)
-                graph.create(movie_to_genre)
+        import_genre_info(graph, genre_ids, movie_node)
+
+        # 导入出品国信息
+        import_country_info(graph, origin_countries, movie_node)
+
+        # 导入出品公司信息
+        import_company_info(graph, companies, movie_node)
+
+        # 导入出品人信息
 
         return "success"
 
@@ -111,13 +168,16 @@ if __name__ == "__main__":
     api_key = arg.key or load_key_from_file("api.key")
     url = f"{arg.host}:{arg.port}"
     graph = Graph(url, username=arg.user, password=arg.password)
-    graph.delete_all()
+    if arg.clearDB:
+        graph.delete_all()
     api = TMDBApi(api_key, arg.language, retry=arg.retry)
 
     # 插入题材节点
-    genre_list = api.get_genre_list(arg.target)
-    for genre in genre_list:
-        insert_if_not_exists(graph, genre["id"], "genre", name=genre["name"])
+    for target in ("movie", "tv"):
+        genre_list = api.get_genre_list(target)
+        for genre in genre_list:
+            insert_if_not_exists(
+                graph, genre["id"], "genre", name=genre["name"])
 
     # 获取top列表
     topN = api.get_popular_iter(arg.target, limit=arg.number)
